MAPFBenchmarks.jl

Docstrings

MAPFBenchmarks.MAPFBenchmarks — Module

MAPFBenchmarks

Interface between the package MultiAgentPathFinding.jl (https://github.com/gdalle/MultiAgentPathFinding.jl) and Sturtevant's MAPF benchmarks (https://movingai.com/benchmarks/mapf.html).

The data files will be downloaded automatically when you run the core function read_benchmark_mapf. See https://movingai.com/benchmarks/formats.html for the description of their format.

MAPFBenchmarks.MAPFBenchmarkProblem — Type

MAPFBenchmarkProblem

Encode one agent of a MAPF scenario.

Fields

index::Int64

bucket::Int64

map_path::String

width::Int64

height::Int64

start_i::Int64

start_j::Int64

goal_i::Int64

goal_j::Int64

optimal_length::Float64

MAPFBenchmarks.benchmark_graph — Method

benchmark_graph(
    map_matrix::Matrix{Char};
    directions=QUEEN_DIRECTIONS_PLUS_CENTER,
    nb_corners_for_diag=2,
    pythagoras_cost_for_diag=true,
)

Create a grid graph from a map specified as a matrix of characters.

Returns a GridGraphs.GridGraph.

MAPFBenchmarks.benchmark_mapf — Method

benchmark_mapf(
    g::GridGraph,
    scenario::Vector{MAPFBenchmarkProblem};
    flexible_departure=false
)

Create a MAPF instance from a grid graph and a scenario (i.e. a set of agents).

Returns a MultiAgentPathFinding.MAPF.

MAPFBenchmarks.cell_color — Method

cell_color(c::Char)

Give a color object corresponding to the type of cell.

To visualize a map in VSCode, just run cell_color.(map_matrix) in the REPL.

MAPFBenchmarks.read_benchmark_map — Method

read_benchmark_map(map_name::AbstractString)

Read a map matrix from a text file.

Returns a Matrix{Char}.

MAPFBenchmarks.read_benchmark_mapf — Method

read_benchmark_mapf(
    map_name::AbstractString,
    scenario_name::AbstractString;
    directions=QUEEN_DIRECTIONS_PLUS_CENTER,
    nb_corners_for_diag=2,
    pythagoras_cost_for_diag=true,
    flexible_departure=false,
)

Create a MAPF instance by reading a map ("something.map") and scenario ("something.scen") from files.

Returns a MultiAgentPathFinding.MAPF.

Example

using Graphs, MAPFBenchmarks, MultiAgentPathFinding
mapf = read_benchmark_mapf("Berlin_1_256.map", "Berlin_1_256-random-1.scen")
nv(mapf.g), nb_agents(mapf)

# output

(65536, 1000)

MAPFBenchmarks.read_benchmark_scenario — Method

read_benchmark_scenario(scenario_name::AbstractString, map_name::AbstractString)

Read a scenario from a text file, and check that it corresponds to a given map.

Returns a Vector{MAPFBenchmarkProblem}.

Index

MAPFBenchmarks.MAPFBenchmarks

MAPFBenchmarks.MAPFBenchmarkProblem

MAPFBenchmarks.benchmark_graph

MAPFBenchmarks.benchmark_mapf

MAPFBenchmarks.cell_color

MAPFBenchmarks.read_benchmark_map

MAPFBenchmarks.read_benchmark_mapf

MAPFBenchmarks.read_benchmark_scenario